+++ app/api/usuario/curriculo/cursos/[id]/route.js
@@ -170,28 +170,29 @@ export async function DELETE(req, { params }) {
         where: { RFC: cert.RFC },
       });
 
-      const cursoNombre =  cert.certificado
-      const directoryPath = path.join(process.cwd(), '/app/resources');
-      const directoryUsrPath = path.join(directoryPath,user.id.toString(),'cursos')
-      const filePath = path.join(directoryUsrPath, cursoNombre);
-
-      try{
-        if (fs.existsSync(filePath)) {
-          // Eliminar el archivo
-          fs.unlink(filePath, (err) => {
-            if (err) {
-              console.error(`Error al eliminar el archivo ${filePath}: ${err}`);
-            } else {
-              console.log(`Archivo ${filePath} eliminado con éxito`);
-            }
-          });
-        } else {
-          console.log(`El archivo ${filePath} no existe`);
+      if(cert.certificado){
+        const cursoNombre =  cert.certificado
+        const directoryPath = path.join(process.cwd(), '/app/resources');
+        const directoryUsrPath = path.join(directoryPath,user.id.toString(),'cursos')
+        const filePath = path.join(directoryUsrPath, cursoNombre);
+        
+        try{
+          if (fs.existsSync(filePath)) {
+            // Eliminar el archivo
+            fs.unlink(filePath, (err) => {
+              if (err) {
+                console.error(`Error al eliminar el archivo ${filePath}: ${err}`);
+              } else {
+                console.log(`Archivo ${filePath} eliminado con éxito`);
+              }
+            });
+          } else {
+            console.log(`El archivo ${filePath} no existe`);
+          }
+        }catch(error){
+          console.log("Ocurrio un error: ", error)
         }
-      }catch(error){
-        console.log("Ocurrio un error: ", error)
       }
-      
     }
     const deletedUser = await prisma.cursos.delete({
       where: { id: parseInt(id) },
